Estimate threshold local projections

Estimate threshold local projections

threshold_LP(
  data,
  regime,
  p = 1,
  horizons = 1,
  freq = "month",
  NW = FALSE,
  NW_lags = NULL,
  NW_prewhite = TRUE
)

Arguments

data
data.frame, matrix, ts, xts, zoo: Endogenous regressors

regime
string: name or regime assignment vector in the design matrix (data)

p
int: lags

horizons
int: forecast horizons

freq
string: frequency of data (day, week, month, quarter, year)

NW
boolean: Newey-West correction on variance-covariance matrix

NW_lags
int: number of lags to use in Newey-West correction

NW_prewhite
boolean: TRUE prewhite option for Newey-West correction (see sandwich::NeweyWest function)

Value

list object with elements data, model, forecasts, residuals; if there is more than one forecast horizon estimated, then model, forecasts, residuals will each be a list where each element corresponds to a single horizon

Examples

if (FALSE) {
threshold_LP(
  data = Data,
  regime = 'regime',
  p = 1,
  horizon = 1,
  freq = 'month')
}