Estimate regime-dependent VAR

Estimate regime-dependent VAR

RVAR(
  data,
  horizon = 10,
  freq = "month",
  type = "const",
  p = 1,
  lag.ic = NULL,
  lag.max = NULL,
  regime = NULL,
  regime.method = "rf",
  regime.n = 2
)

Arguments

data
data.frame, matrix, ts, xts, zoo: Endogenous regressors

horizon
int: forecast horizons

freq
string: frequency of data ('day', 'week', 'month', 'quarter', or 'year')

type
string: type of deterministic terms to add ('none', 'const', 'trend', or 'both')

p
int: lags

lag.ic
string: information criterion to choose the optimal number of lags ('AIC' or 'BIC')

lag.max
int: maximum number of lags to test in lag selection

regime
string: name or regime assignment vector in the design matrix (data)

regime.method
string: regime assignment technique ('rf', 'kmeans', 'EM', or 'BP')

regime.n
int: number of regimes to estimate (applies to kmeans and EM)

Value

list of lists, each regime returns its own list with elements data, model, forecasts, residuals

See also

Examples

# \donttest{

 # simple time series
 AA = c(1:100) + rnorm(100)
 BB = c(1:100) + rnorm(100)
 CC = AA + BB + rnorm(100)
 date = seq.Date(from = as.Date('2000-01-01'), by = 'month', length.out = 100)
 Data = data.frame(date = date, AA, BB, CC)
 Data = dplyr::mutate(Data, reg = dplyr::if_else(AA > median(AA), 1, 0))

 # estimate VAR
  rvar =
    sovereign::RVAR(
      data = Data,
      horizon = 10,
      freq = 'month',
      regime.method = 'rf',
      regime.n = 2,
      lag.ic = 'BIC',
      lag.max = 4)

#> Warning: NAs introduced by coercion

#> Warning: NAs introduced by coercion

#> Warning: NAs introduced by coercion

#> Warning: NAs introduced by coercion

# impulse response functions
rvar.irf = sovereign::rvar_irf(rvar)

# forecast error variance decomposition
rvar.fevd = sovereign::rvar_fevd(rvar)

# }