Estimate multi-regime local projections

Estimate multi-regime local projections

threshold_LP(
  data,
  horizons = 1,
  freq = "month",
  type = "const",
  p = 1,
  lag.ic = NULL,
  lag.max = NULL,
  NW = FALSE,
  NW_lags = NULL,
  NW_prewhite = NULL,
  regime = NULL,
  regime.method = "rf",
  regime.n = 2
)

Arguments

data
data.frame, matrix, ts, xts, zoo: Endogenous regressors

horizons
int: forecast horizons

freq
string: frequency of data (day, week, month, quarter, year)

type
string: type of deterministic terms to add ('none', 'const', 'trend', 'both')

p
int: lags

lag.ic
string: information criterion to choose the optimal number of lags ('AIC' or 'BIC')

lag.max
int: maximum number of lags to test in lag selection

NW
boolean: Newey-West correction on variance-covariance matrix

NW_lags
int: number of lags to use in Newey-West correction

NW_prewhite
boolean: TRUE prewhite option for Newey-West correction (see sandwich::NeweyWest function)

regime
string: name or regime assignment vector in the design matrix (data)

regime.method
string: regime assignment technique ('rf', 'kmeans', 'EM', 'BP')

regime.n
int: number of regimes to estimate (applies to kmeans and EM)

Value

list object with elements data, model, forecasts, residuals; if there is more than one forecast horizon estimated, then model, forecasts, residuals will each be a list where each element corresponds to a single horizon

Examples

# \donttest{

  # simple time series
  AA = c(1:100) + rnorm(100)
  BB = c(1:100) + rnorm(100)
  CC = AA + BB + rnorm(100)
  date = seq.Date(from = as.Date('2000-01-01'), by = 'month', length.out = 100)
  Data = data.frame(date = date, AA, BB, CC)
  # add regime
  Data = dplyr::mutate(Data, reg = dplyr::if_else(AA > median(AA), 1, 0))

  # local projection forecasts
  tlp =
    threshold_LP(
      data = Data,
      regime = 'reg',
      horizon = c(1:10),
      freq = 'month',
      p = 1,
      type =  'const',
      NW = TRUE,
      NW_lags = 1,
      NW_prewhite = FALSE)

 # impulse response function
 tirf = threshold_lp_irf(tlp)

# }